requirements.txt
fastapi
motor
passlib
pydantic
python-dotenv
uvicorn
